requirements.txt
boto3==1.17.78
botocore==1.20.78
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
